Setting workspace

+++ src/App.js
@@ -1,22 +1,21 @@
-
 import Workspace from "./Workspace/Workspace";
 import Sidebar from "./Sidebar/Sidebar.jsx";
 import SearchBox from "./SearchBox/SearchBox.jsx";
 import { db } from "./db";
 import { useLiveQuery } from "dexie-react-hooks";
 import NotesList from "./NotesList/NotesList";
-import { createContext } from 'react';
+import { createContext, useState } from "react";
 
 export const NotesContext = createContext({});
 
 const App = (props) => {
-  console.log(props)
+  const [activeNote, getActiveNote] = useState("");
+  console.log(props);
   let isFinished = false;
-  async function addNote(inputTextNote) {
+  async function addNote(note) {
     try {
-      // Add the new friend!
       const id = await db.notes.add({
-        inputTextNote,
+        note,
         createdAt: new Date(),
       });
     } catch (error) {
@@ -25,34 +24,69 @@ const App = (props) => {
   }
   const notesList = useLiveQuery(() => db.notes.toArray());
 
-  
-  let noteToDel
+  let noteToDel = "222222222";
 
-function giveNoteToDelete(note){
-  console.log(note)
-  noteToDel=note;
-  console.log(noteToDel)
-  return noteToDel;
-}
-function deleteNote(){
-    
-  db.notes.where('id').equals(noteToDel.id).delete()
+  function giveNoteToDelete(note) {
+    console.log(note);
+    noteToDel = note;
+    console.log(noteToDel);
+    getActiveNote(noteToDel.note);
+    return noteToDel;
+  }
 
+  function deleteNote() {
+    let confirming = window.confirm("Delete note?");
+    if (confirming) {
+      db.notes.where("id").equals(noteToDel.id).delete();
     }
+  }
+  console.log(noteToDel);
+
+  function changeNote() {}
+
+  async function toSearch(searchText) {
+    console.log(db.notes);
+
+    const result = db.notes
+      .filter(function (note) {
+        return /dd/.test(note);
+      })
+      .toArray();
+    console.log(result);
+  }
+  let isActive = false;
+  function changeNote() {
+    isActive = true;
+  }
 
   return (
     <>
-    <NotesContext.Provider value={{deleteNote, addNote,notesList}}>
-      <div className="App">
-        <div className="navbar">
-          <Sidebar addNote={addNote} deleteNote={deleteNote} />
-          <SearchBox />
+      <NotesContext.Provider
+        value={{
+          deleteNote,
+          addNote,
+          notesList,
+          giveNoteToDelete,
+          noteToDel,
+          activeNote,
+          changeNote,
+          isActive,
+        }}
+      >
+        <div className="App">
+          <div className="navbar">
+            <Sidebar addNote={addNote} deleteNote={deleteNote} />
+            <SearchBox notesList={notesList} toSearch={toSearch} />
+          </div>
+          <div className="container">
+            <NotesList
+              notesList={notesList}
+              isFinished={isFinished}
+              noteToDelete={giveNoteToDelete}
+            />
+            <Workspace activeNote={activeNote} />
+          </div>
         </div>
-        <div className="container">
-          <NotesList notesList={notesList} isFinished={isFinished} noteToDelete={giveNoteToDelete}/>
-          <Workspace />
-        </div>
-      </div>
       </NotesContext.Provider>
     </>
   );
